Add `RBS::UnitTest` module
