+++ lib/rbs/test.rb
@@ -3,7 +3,6 @@
 require "securerandom"
 require "rbs/test/guaranteed"
 require "rbs/test/observer"
-require "rbs/test/spy"
 require "rbs/test/errors"
 require "rbs/test/type_check"
 require "rbs/test/tester"
@@ -85,11 +84,11 @@ module RBS
 
     if ::UnboundMethod.instance_methods.include?(:bind_call)
       def self.call(receiver, method, *args, &block)
-        method.bind_call(receiver, *args, &block)
+        __skip__ = method.bind_call(receiver, *args, &block)
       end
     else
       def self.call(receiver, method, *args, &block)
-        method.bind(receiver).call(*args, &block)
+        __skip__ = method.bind(receiver).call(*args, &block)
       end
     end
   end
