--- lib/rbs/test/spy.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module RBS
-  module Test
-  end
-end
