+++ test/stdlib/test_helper.rb
@@ -21,124 +21,6 @@ class Test::Unit::TestCase
   # prepend Printer
 end
 
-module Spy
-  def self.wrap(object, method_name)
-    spy = WrapSpy.new(object: object, method_name: method_name)
-
-    if block_given?
-      begin
-        yield spy, spy.wrapped_object
-      end
-    else
-      spy
-    end
-  end
-
-  class WrapSpy
-    attr_accessor :callback
-    attr_reader :object
-    attr_reader :method_name
-
-    def initialize(object:, method_name:)
-      @callback = -> (_) { }
-      @object = object
-      @method_name = method_name
-    end
-
-    def wrapped_object
-      spy = self
-
-      Class.new(BasicObject) do
-        define_method(:method_missing) do |name, *args, &block|
-          spy.object.__send__(name, *args, &block)
-        end
-
-        define_method(spy.method_name, -> (*args, &block) {
-          return_value = nil
-          exception = nil
-          block_calls = []
-
-          spy_block = if block
-                        Object.new.instance_eval do |fresh|
-                          -> (*block_args) do
-                            block_exn = nil
-                            block_return = nil
-
-                            begin
-                              block_return = if self.equal?(fresh)
-                                               # no instance eval
-                                               block.call(*block_args)
-                                             else
-                                               self.instance_exec(*block_args, &block)
-                                             end
-                            rescue Exception => exn
-                              block_exn = exn
-                            end
-
-                            if block_exn
-                              block_calls << RBS::Test::ArgumentsReturn.exception(
-                                arguments: block_args,
-                                exception: block_exn
-                              )
-                            else
-                              block_calls << RBS::Test::ArgumentsReturn.return(
-                                arguments: block_args,
-                                value: block_return
-                              )
-                            end
-
-                            if block_exn
-                              raise block_exn
-                            else
-                              block_return
-                            end
-                          end.ruby2_keywords
-                        end
-                      end
-
-          begin
-            if spy_block
-              return_value = spy.object.__send__(spy.method_name, *args) do |*a, **k, &b|
-                spy_block.call(*a, **k, &b)
-              end
-            else
-              return_value = spy.object.__send__(spy.method_name, *args, &spy_block)
-            end
-          rescue ::Exception => exn
-            exception = exn
-          end
-
-          call = if exception
-                   RBS::Test::ArgumentsReturn.exception(
-                     arguments: args,
-                     exception: exception
-                   )
-                 else
-                   RBS::Test::ArgumentsReturn.return(
-                     arguments: args,
-                     value: return_value
-                   )
-                 end
-          trace = RBS::Test::CallTrace.new(
-            method_name: spy.method_name,
-            method_call: call,
-            block_calls: block_calls,
-            block_given: block != nil
-          )
-
-          spy.callback.call(trace)
-
-          if exception
-            spy.object.__send__(:raise, exception)
-          else
-            return_value
-          end
-        }.ruby2_keywords)
-      end.new()
-    end
-  end
-end
-
 module VersionHelper
   def if_ruby(range)
     r = Range.new(
@@ -300,303 +182,6 @@ module WithAliases
         yield Range.new(lower, upper, exclude_end)
       end
     end
-  end
-end
-
-module TypeAssertions
-  module ClassMethods
-    attr_reader :target
-
-    def library(*libs)
-      if @libs
-        raise "Multiple #library calls are not allowed"
-      end
-
-      @libs = libs
-      @env = nil
-      @target = nil
-    end
-
-    @@env_cache = {}
-
-    def env
-      @env = @@env_cache[@libs] ||=
-        begin
-          loader = RBS::EnvironmentLoader.new
-          (@libs || []).each do |lib|
-            loader.add library: lib
-          end
-
-          RBS::Environment.from_loader(loader).resolve_type_names
-        end
-    end
-
-    def builder
-      @builder ||= RBS::DefinitionBuilder.new(env: env)
-    end
-
-    def testing(type_or_string)
-      type = case type_or_string
-             when String
-               RBS::Parser.parse_type(type_or_string, variables: [])
-             else
-               type_or_string
-             end
-
-      definition = case type
-                   when RBS::Types::ClassInstance
-                     builder.build_instance(type.name)
-                   when RBS::Types::ClassSingleton
-                     builder.build_singleton(type.name)
-                   else
-                     raise "Test target should be class instance or class singleton: #{type}"
-                   end
-
-      @target = [type, definition]
-    end
-  end
-
-  def self.included(base)
-    base.extend ClassMethods
-  end
-
-  def env
-    self.class.env
-  end
-
-  def builder
-    self.class.builder
-  end
-
-  def targets
-    @targets ||= []
-  end
-
-  def target
-    targets.last || self.class.target
-  end
-
-  def testing(type_or_string)
-    type = case type_or_string
-           when String
-             RBS::Parser.parse_type(type_or_string, variables: [])
-           else
-             type_or_string
-           end
-
-    definition = case type
-                 when RBS::Types::ClassInstance
-                   builder.build_instance(type.name)
-                 when RBS::Types::ClassSingleton
-                   builder.build_singleton(type.name)
-                 else
-                   raise "Test target should be class instance or class singleton: #{type}"
-                 end
-
-    targets.push [type, definition]
-
-    if block_given?
-      begin
-        yield
-      ensure
-        targets.pop
-      end
-    else
-      [type, definition]
-    end
-  end
-
-  def instance_class
-    type, _ = target
-
-    case type
-    when RBS::Types::ClassSingleton, RBS::Types::ClassInstance
-      Object.const_get(type.name.to_s)
-    end
-  end
-
-  def class_class
-    type, _ = target
-
-    case type
-    when RBS::Types::ClassSingleton, RBS::Types::ClassInstance
-      Object.const_get(type.name.to_s).singleton_class
-    end
-  end
-
-  ruby2_keywords def assert_send_type(method_type, receiver, method, *args, &block)
-    trace = []
-    spy = Spy.wrap(receiver, method)
-    spy.callback = -> (result) { trace << result }
-
-    result = nil
-    exception = nil
-
-    begin
-      result = spy.wrapped_object.__send__(method, *args, &block)
-    rescue Exception => exn
-      exception = exn
-    end
-
-    mt = case method_type
-         when String
-           RBS::Parser.parse_method_type(method_type, variables: [])
-         when RBS::MethodType
-           method_type
-         end
-
-    typecheck = RBS::Test::TypeCheck.new(
-      self_class: receiver.class,
-      builder: builder,
-      sample_size: 100,
-      unchecked_classes: [],
-      instance_class: instance_class,
-      class_class: class_class
-    )
-    errors = typecheck.method_call(method, mt, trace.last, errors: [])
-
-    assert_empty errors.map {|x| RBS::Test::Errors.to_string(x) }, "Call trace does not match with given method type: #{trace.last.inspect}"
-
-    method_types = method_types(method)
-    all_errors = method_types.map {|t| typecheck.method_call(method, t, trace.last, errors: []) }
-    assert all_errors.any? {|es| es.empty? }, "Call trace does not match one of method definitions:\n  #{trace.last.inspect}\n  #{method_types.join(" | ")}"
-
-    raise exception if exception
-
-    result
-  end
-
-  ruby2_keywords def refute_send_type(method_type, receiver, method, *args, &block)
-    trace = []
-    spy = Spy.wrap(receiver, method)
-    spy.callback = -> (result) { trace << result }
-
-    result = nil
-    exception = nil
-
-    begin
-      result = spy.wrapped_object.__send__(method, *args, &block)
-    rescue Exception => exn
-      exception = exn
-    end
-
-    mt = case method_type
-         when String
-           RBS::Parser.parse_method_type(method_type, variables: [])
-         when RBS::MethodType
-           method_type
-         end
-
-    mt = mt.update(block: if mt.block
-                            RBS::Types::Block.new(
-                              type: mt.block.type.with_return_type(RBS::Types::Bases::Any.new(location: nil)),
-                              required: mt.block.required,
-                              self_type: nil
-                            )
-                          end,
-                   type: mt.type.with_return_type(RBS::Types::Bases::Any.new(location: nil)))
-
-    typecheck = RBS::Test::TypeCheck.new(
-      self_class: receiver.class,
-      instance_class: instance_class,
-      class_class: class_class,
-      builder: builder,
-      sample_size: 100,
-      unchecked_classes: []
-    )
-    errors = typecheck.method_call(method, mt, trace.last, errors: [])
-
-    assert_operator exception, :is_a?, ::Exception
-    assert_empty errors.map {|x| RBS::Test::Errors.to_string(x) }
-
-    method_types = method_types(method)
-    all_errors = method_types.map {|t| typecheck.method_call(method, t, trace.last, errors: []) }
-    assert all_errors.all? {|es| es.size > 0 }, "Call trace unexpectedly matches one of method definitions:\n  #{trace.last.inspect}\n  #{method_types.join(" | ")}"
-
-    result
-  end
-
-  def method_types(method)
-    type, definition = target
-
-    case
-    when definition.instance_type?
-      subst = RBS::Substitution.build(definition.type_params, type.args)
-      definition.methods[method].method_types.map do |method_type|
-        method_type.sub(subst)
-      end
-    when definition.class_type?
-      definition.methods[method].method_types
-    end
-  end
-
-  def allows_error(*errors)
-    yield
-  rescue *errors => exn
-    notify "Error allowed: #{exn.inspect}"
-  end
-
-  include VersionHelper
-  include WithAliases
-
-  def assert_const_type(type, constant_name)
-    constant = Object.const_get(constant_name)
-
-    typecheck = RBS::Test::TypeCheck.new(
-      self_class: constant.class,
-      instance_class: instance_class,
-      class_class: class_class,
-      builder: builder,
-      sample_size: 100,
-      unchecked_classes: []
-    )
-
-    value_type =
-      case type
-      when String
-        RBS::Parser.parse_type(type, variables: [])
-      else
-        type
-      end
-
-    assert typecheck.value(constant, value_type), "`#{constant_name}` (#{constant.inspect}) must be compatible with given type `#{value_type}`"
-
-    type_name = TypeName(constant_name).absolute!
-    definition = env.constant_entry(type_name)
-    assert definition, "Cannot find RBS type definition of `#{constant_name}`"
-
-    case definition
-    when RBS::Environment::ClassEntry, RBS::Environment::ModuleEntry
-      definition_type = RBS::Types::ClassSingleton.new(name: type_name, location: nil)
-    when RBS::Environment::ClassAliasEntry, RBS::Environment::ModuleAliasEntry
-      type_name = env.normalize_type_name!(type_name)
-      definition_type = RBS::Types::ClassSingleton.new(name: type_name, location: nil)
-    when RBS::Environment::ConstantEntry
-      definition_type = definition.decl.type
-    end
-
-    assert typecheck.value(constant, definition_type), "`#{constant_name}` (#{constant.inspect}) must be compatible with RBS type definition `#{definition_type}`"
-  end
-
-  def assert_type(type, value)
-    typecheck = RBS::Test::TypeCheck.new(
-      self_class: value.class,
-      instance_class: "No `instance` class allowed",
-      class_class: "No `class` class allowed",
-      builder: builder,
-      sample_size: 100,
-      unchecked_classes: []
-    )
-
-    type =
-      case type
-      when String
-        RBS::Parser.parse_type(type, variables: [])
-      else
-        type
-      end
-
-    assert typecheck.value(value, type), "`#{value.inspect}` must be compatible with given type `#{type}`"
   end
 end
